% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_daynight.r
\name{add_daynight}
\alias{add_daynight}
\title{Add day night designation to series}
\usage{
add_daynight(series, etuff)
}
\arguments{
\item{series}{is input time series data}

\item{etuff}{is etuff object}
}
\value{
a vector of day/night labels that coincides with input time series
}
\description{
Add day night designation to series based on local sunrise sunset times
}
