% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_series_temp.r
\name{add_series_temp}
\alias{add_series_temp}
\title{Add temperature to series data, if possible}
\usage{
add_series_temp(series, pdt = NULL, pdt_interp = NULL, flag = TRUE)
}
\arguments{
\item{series}{is output from \link{\code{get_series}}}

\item{pdt}{is output from \link{\code{get_pdt}}}

\item{pdt_interp}{is output from \link{\code{interp_pdt}}}

\item{flag}{is logical indicating whether or not to flag potentially bad interpolated points. This should almost always be TRUE (the default).}
}
\value{
a new series dataframe with additional temperature values where appropriate
}
\description{
Add missing temperature values to series data from interpolated depth-temperature profile data
}
