% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/archival_to_etuff.r
\name{archival_to_etuff}
\alias{archival_to_etuff}
\title{Convert recovered tag to eTUFF}
\usage{
archival_to_etuff(archival, vars)
}
\arguments{
\item{archival}{is long data frame of archival data}

\item{vars}{is character list of variables of interest. These must match accepted variable names (see obsTypes).}
}
\value{
a dataframe subset of the archival data containing only the vars specified
}
\description{
Converts recovered or archival tag dataset to eTUFF in a "computationally-friendly" way compared to the normal conversion.
}
