% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.light.R
\name{extract.light}
\alias{extract.light}
\title{Extract lgiht from Wildlife Computers tag data}
\usage{
extract.light(light)
}
\arguments{
\item{lightloc}{data frame read from -LightLoc.csv output of Wildlife Computers DAP
processor or Tag Portal.}
}
\value{
data frame formatted for LL data
}
\description{
\code{extract.light} is a simple formatting function that parses LL data and
makes it usable to subsequent functions
}
