% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findDateFormat.r
\name{findDateFormat}
\alias{findDateFormat}
\title{Determine date format of vector}
\usage{
findDateFormat(dateVec)
}
\arguments{
\item{dateVec}{a character vector representing dates}
}
\value{
dateformat is character string used as input to strptime(format = dateformat)
}
\description{
\code{findDateFormat} determines the date format of a given vector of dates
}
\examples{
dte <- '2015-01-01 05:30:17'
findDateFormat(dte)
dte.POSIX <- as.POSIXct(dte, format = findDateFormat(dte))
dte.POSIX

}
