% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_3d.r
\name{get_3d}
\alias{get_3d}
\title{Get 3d data from eTUFF}
\usage{
get_3d(etuff, series = NULL, fG = TRUE, ...)
}
\arguments{
\item{etuff}{is object of class etuff}

\item{series}{is optional series as input. Typically this is NULL but can be supplied to speed up processing (i.e. if series has already been processed).}

\item{fG}{is logical indicating whether or not to use foieGras to interpolate and standardize tracking data. Note this may already be done prior to building the eTUFF file and thus may be unncessary. See the eTUFF object's metadata/header info.}
}
\value{
a dataframe of 3d movements
}
\description{
Extract location and vertical time series information and combine them into a dataframe that represents 3d movements.
}
