% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_etuff_hdr.r
\name{get_etuff_hdr}
\alias{get_etuff_hdr}
\title{Get header as metadata from eTUFF file}
\usage{
get_etuff_hdr(etuff_file)
}
\arguments{
\item{etuff_file}{is file name to etuff file of interest}
}
\value{
a dataframe of metadata from the header of an eTUFF file
}
\description{
WARNING: this is an internal function. Instead see \code{get_header}.
}
