% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_header.r
\name{get_header}
\alias{get_header}
\title{Get header as metadata from eTUFF file}
\usage{
get_header(etuff_file, metaTypes = NULL)
}
\arguments{
\item{etuff_file}{is file name to etuff file of interest}

\item{metaTypes}{is dataframe of allowed metadata types from github}
}
\value{
a dataframe of metadata from the header of an eTUFF file
}
\description{
Get header as metadata from eTUFF file
}
