% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_mmd.r
\name{get_mmd}
\alias{get_mmd}
\title{Get min/max depth data from eTUFF}
\usage{
get_mmd(etuff)
}
\arguments{
\item{etuff}{is object of class etuff}
}
\value{
a dataframe of min/max depth information
}
\description{
Extract min/max depth data as typically contained in -MinMaxDepth.csv output from Wildlife Computers
}
