% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_pdt.r
\name{get_pdt}
\alias{get_pdt}
\title{Get depth-temperature profile data from eTUFF}
\usage{
get_pdt(etuff)
}
\arguments{
\item{etuff}{is object of class etuff}
}
\value{
a dataframe of min/max depth information
}
\description{
Extract depth-temperature profile data as typically contained in -PDTs.csv output from Wildlife Computers
}
