% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_series.r
\name{get_series}
\alias{get_series}
\title{Extract time series data as typically contained in -Series.csv output from Wildlife Computers}
\usage{
get_series(etuff, temp_res = NULL, what_tz = NULL)
}
\arguments{
\item{etuff}{is object of class etuff}

\item{temp_res}{is temporal resolution as valid argument to "by" in \code{seq.POSIXct}}

\item{what_tz}{indicates what time zone should be used for the series data. If NULL (default), the function will choose for you with some informative printed outputs to let you know whats happening.}
}
\value{
a dataframe of time series data, usually containing at least depth and temperature information
}
\description{
Extract time series data as typically contained in -Series.csv output from Wildlife Computers
}
