% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_srss.r
\name{get_srss}
\alias{get_srss}
\title{Get sunrise/sunset times}
\usage{
get_srss(etuff)
}
\arguments{
\item{etuff}{is object of class etuff}
}
\value{
a dataframe of sunrise and sunset information
}
\description{
Get sunrise/sunset times concurrent to the timeseries data in eTUFF. Typically this is used to label time series information with relevant SRSS times.
}
