% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_tad.r
\name{get_tad}
\alias{get_tad}
\title{Get time-at-depth data from eTUFF}
\usage{
get_tad(etuff)
}
\arguments{
\item{etuff}{is object of class etuff}
}
\value{
a dataframe of min/max depth information
}
\description{
Extract time-at-depth data as typically contained in -Histos.csv output from Wildlife Computers
}
