% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_tat.r
\name{get_tat}
\alias{get_tat}
\title{Get time-at-temperature data from eTUFF}
\usage{
get_tat(etuff)
}
\arguments{
\item{etuff}{is object of class etuff}
}
\value{
a dataframe of min/max depth information
}
\description{
Extract time-at-temperature data as typically contained in -Histos.csv output from Wildlife Computers
}
