% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_track.r
\name{get_track}
\alias{get_track}
\title{Get track data from eTUFF}
\usage{
get_track(etuff, what_tz = NULL)
}
\arguments{
\item{etuff}{is object of class etuff}

\item{what_tz}{indicates the timezone (usually local time) for this tag dataset. If NULL (default), the function will use the timezone from the tagging location. In the future, this will get more sophisticated by attempting to query tz at each timestamp for associated lat/lon.}
}
\value{
a dataframe of lat/lon and (sometimes) associated error
}
\description{
Extract track data from eTUFF
}
