% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_tz.r
\name{get_tz}
\alias{get_tz}
\title{Get time zone(s) associated with the eTUFF object of interest}
\usage{
get_tz(etuff, what_tz = NULL)
}
\arguments{
\item{etuff}{is object of class etuff}

\item{what_tz}{is optional specification for time zone to use and bypass the search for one}
}
\value{
a vector containing the time zone(s) relevant to the eTUFF object of interest
}
\description{
Get time zone(s) associated with the eTUFF object of interest
}
