% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCtr_gpe3.r
\name{getCtr_gpe3}
\alias{getCtr_gpe3}
\title{Get specified contour coordinates}
\usage{
getCtr_gpe3(ncFile, csvFile, threshold = 50, makePlot = FALSE)
}
\arguments{
\item{ncFile}{is output nc from GPE3}

\item{threshold}{numeric indicating the percent contour of interest. Default
is 50 percent.}

\item{makePlot}{is logical indicating whether or not to plot at each iteration}
}
\value{
a list of length T, dim(distr)[2], containing 1) coordinates of the
  contour at each time, t. 2) the x and y distance to that contour from the
  mean of the distribution (lat/lon in track). 3) reference coordinate from
  mean of distribution. 4) the points of intersection of the contour by which
  distance is measured in x and y
}
\description{
\code{getCtr} uses GPE3 12 hour likelihood output to calculate coordinates at a
specified contour of each posterior distribution of the state
}
