% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSeriesTemp.r
\name{getSeriesTemp}
\alias{getSeriesTemp}
\title{Extract relevant temperatures from interpolated depth-temperature data}
\usage{
getSeriesTemp(series, pdt, loess, flag = TRUE)
}
\arguments{
\item{series}{is data frame of Series data unformatted from WC Portal
download}

\item{pdt}{is data frame of PDT data formatted and output from
\code{prep.loess}}

\item{loess}{is list output from \code{grid2dloess}}

\item{flag}{is logical indicating whether or not to mask out flagged values
in LOESS output. Default is TRUE in which flagged values are masked out.}
}
\value{
a data frame of series data with (hopefully) added temperature data from the interpolated temperatures
}
\description{
Extract relevant temperatures from interpolated depth-temperature data. This is used to augment existing time series temperature data (if any).
}
