% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interp_pdt.r
\name{interp_pdt}
\alias{interp_pdt}
\title{Interpolate depth-temperature profile data to gap-fill}
\usage{
interp_pdt(pdt, span_x = 5, span_y = 150)
}
\arguments{
\item{pdt}{is dataframe of depth-temperature profile data}

\item{span_x}{is interpolation input param}

\item{span_y}{is interpolation input param}
}
\value{
a list containing a data frame of interpolated depth-temperature data and a matching data frame with flags to highlihgt potential interpolation issues
}
\description{
Interpolate depth-temperature profile data to gap-fill
}
