% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mode.r
\name{Mode}
\alias{Mode}
\title{Compute the mode of an input vector}
\usage{
Mode(x, na.rm = FALSE)
}
\arguments{
\item{x}{is input vector}
}
\value{
statistical mode of input vector
}
\description{
Compute the mode of an input vector
}
