% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_sst.r
\name{parse_sst}
\alias{parse_sst}
\title{Prepare SST data from mixed sources for eTUFF standardization}
\usage{
parse_sst(sst, obsTypes)
}
\arguments{
\item{sst}{is data frame of data from Wildlife Computers -SST.csv}

\item{obsTypes}{is data frame of observation types read from Github}
}
\value{
a "master" sst dataset compiled across potential sources
}
\description{
Prepare SST data from mixed sources for eTUFF standardization. This is nearly always useful for dealing with Wildlife Computers tag datasets when SSTs can be derived from multiple "sources".
}
