% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_archival_etuff.r
\name{qc_archival_etuff}
\alias{qc_archival_etuff}
\title{Simple quality control plots for archival eTUFF file}
\usage{
qc_archival_etuff(etuff, meta_row, writePNG = FALSE, map = TRUE)
}
\arguments{
\item{etuff}{is etuff file}

\item{meta_row}{is the associated metadata}

\item{writePNG}{is logical indicating whether or not to write the output to file as PNG}

\item{map}{is logical indicating whether to try to add a map (assuming track data is present in eTUFF)}
}
\value{
a QC plot
}
\description{
This function is to do very simple QC on PSAT tag conversion to eTUFF and the
associated metadata. It is meant to check that the converted data make sense.
}
