% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_spot_etuff.r
\name{qc_spot_etuff}
\alias{qc_spot_etuff}
\title{Simple quality control plots for SPOT eTUFF file}
\usage{
qc_spot_etuff(etuff, meta_row, writePNG = FALSE)
}
\arguments{
\item{etuff}{is etuff file}

\item{meta_row}{is the associated metadata}

\item{writePNG}{is logical indicating whether or not to write the output to file as PNG}
}
\value{
a QC plot
}
\description{
This function is to do very simple QC on SPOT tag conversion to eTUFF and the
associated metadata. It does NOT do any filtering or further processing of
the positions. Instead, it is meant to check that the coordinates, time
stamps, location classes and associated metadata all make sense.
}
