% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_archival.r
\name{read_archival}
\alias{read_archival}
\title{Read an archival tag stored as eTUFF}
\usage{
read_archival(etuff_file, header = TRUE, metaTypes = NULL)
}
\arguments{
\item{etuff_file}{is an etuff text file}

\item{header}{is logical indicating whether or not the target etuff_file has a header. This will nearly always be TRUE (default).}

\item{metaTypes}{is a dataframe that describes the appropriate inventory of metadata vocabulary. Default is NULL in which this table is read from Github.}
}
\value{
an etuff_archival object
}
\description{
Read an archival tag stored as eTUFF. This only differs from \code{read_etuff} in that it cuts some corners in an attempt to read such a large file faster.
}
