% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_etuff.r
\name{read_etuff}
\alias{read_etuff}
\title{Read an eTUFF file}
\usage{
read_etuff(etuff_file, header = TRUE, metaTypes = NULL)
}
\arguments{
\item{etuff_file}{is an etuff text file}

\item{header}{is logical indicating whether or not the target etuff_file has a header. This will nearly always be TRUE (default).}

\item{metaTypes}{is a dataframe that describes the appropriate inventory of metadata vocabulary. Default is NULL in which this table is read from Github.}
}
\value{
an etuff object
}
\description{
Read tag data stored as eTUFF
}
