% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tag_to_etuff.r
\name{tag_to_etuff}
\alias{tag_to_etuff}
\title{Convert satellite tag data to etuff}
\usage{
tag_to_etuff(dir, meta_row, fName = NULL, tatBins = NULL,
  tadBins = NULL, obsTypes = NULL, check_meta = TRUE, ...)
}
\arguments{
\item{dir}{is directory the target data is stored in}

\item{fName}{is character indicating the file name of interest to read. This
is currently only required for reading archival data from Microwave or Lotek
tags. It can also be used to read individual files from WC or custom inputs. See `customCols` parameter.}

\item{tatBins}{is integer or numeric vector indicating the
time-at-temperature summary bins for a Wildlife tag. Length will usually be
about 14-16 bins. Defaults to NULL and the function tries to read bins from
the file.}

\item{tadBins}{is integer or numeric vector indicating the time-at-depth
summary bins for a Wildlife tag. Length will usually be about 14-16 bins.
Defaults to NULL and the function tries to read bins from the file.}

\item{obsTypes}{is csv sourced from github containing the latest obsTypes
recognized by the NASA OIIP project. Usually this is left NULL and the file
is auto-magically downloaded for you. The only reason you may want to
specify this would be in order to work offline.}

\item{check_meta}{is logical indicating whether or not to check the etuff file metadata}

\item{manufacturer}{is character indicating tag manufacturer. Choices are
'Wildlife','Microwave','Lotek'.}

\item{tagtype}{is character. Choices are 'PSAT', 'SPOT', ...}

\item{dates}{is POSIXct vector of length 2 indicating start and stop dates
for the tag data of interest}

\item{customCols}{is optional argument that allows custom specification of input columns for input \code{fName}. these custom specs must match the accepted obsTypes}
}
\description{
Convert satellite tag data to etuff. Currently this function has good support for Wildlife Computers tags and is in development for others such as Lotek and MT.
}
