% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testDates.r
\name{testDates}
\alias{testDates}
\title{Try to figure out the date format}
\usage{
testDates(x)
}
\arguments{
\item{x}{is input character vector of dates}
}
\value{
a POSIXct vector of dates
}
\description{
Try to figure out the date format and coerce to something useful
}
