% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_etuff.r
\name{write_etuff}
\alias{write_etuff}
\title{Write eTUFF output file}
\usage{
write_etuff(etuff, meta_row = NULL, etuff_file, check_meta = TRUE, ...)
}
\arguments{
\item{etuff}{input etuff object}

\item{meta_row}{optional metadata input but typically the etuff file already has this}

\item{etuff_file}{is desired output etuff file}

\item{check_meta}{is logical indicating whether the metadata should be checked prior to writing}

\item{metaTypes}{optional. is a dataframe that describes the appropriate inventory of metadata vocabulary. Default is NULL in which this table is read from Github.}
}
\value{
writes output etuff to disk
}
\description{
Write eTUFF output file
}
